+++ wai-app-static/ChangeLog.md
@@ -1,3 +1,7 @@
+## 3.0.0.6
+
+Fix trailing slashes for `UrlMap` and other non-root setups [#325](https://github.com/yesodweb/wai/issues/325)
+
 ## 3.0.0.4
 
 Add missing trailing slashes [#312](https://github.com/yesodweb/wai/issues/312)