+++ wai-extra/ChangeLog.md
@@ -1,3 +1,7 @@
+## 3.0.4.2
+
+* `UrlMap`: do not modify `rawPathInfo`, see [#325](https://github.com/yesodweb/wai/issues/325)
+
 ## 3.0.4.1
 
 Fix compilation failure on Windows [#321](https://github.com/yesodweb/wai/issues/321)